requirements.txt
numpy
matplotlib
scipy
ffmpeg-python
PyYAML